separate test_simple.py
